requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-amazon
boto3
requests
psycopg2-binary
pandas
sqlalchemy
pandas-profiling
watchtower
prometheus_client
great_expectations
slack-sdk
msal
